+++ src/main/java/Idea/Archive/IdeaArchive/domain/auth/service/GithubAuthService.java
@@ -6,9 +6,8 @@ import Idea.Archive.IdeaArchive.domain.auth.repository.RefreshTokenRepository;
 import Idea.Archive.IdeaArchive.domain.member.entity.Member;
 import Idea.Archive.IdeaArchive.domain.member.repository.MemberRepository;
 import Idea.Archive.IdeaArchive.global.filter.role.Role;
-import Idea.Archive.IdeaArchive.global.security.GithubAuthProperties;
+import Idea.Archive.IdeaArchive.global.oauth.GithubAuthProperties;
 import Idea.Archive.IdeaArchive.global.security.jwt.TokenProvider;
-import Idea.Archive.IdeaArchive.global.security.jwt.properties.JwtProperties;
 import Idea.Archive.IdeaArchive.infrastructure.feign.client.GithubAuth;
 import Idea.Archive.IdeaArchive.infrastructure.feign.client.GithubEmailInfo;
 import Idea.Archive.IdeaArchive.infrastructure.feign.client.GithubNameInfo;
