+++ src/main/java/Idea/Archive/IdeaArchive/domain/auth/service/GoogleAuthService.java
@@ -6,7 +6,7 @@ import Idea.Archive.IdeaArchive.domain.auth.repository.RefreshTokenRepository;
 import Idea.Archive.IdeaArchive.domain.member.entity.Member;
 import Idea.Archive.IdeaArchive.domain.member.repository.MemberRepository;
 import Idea.Archive.IdeaArchive.global.filter.role.Role;
-import Idea.Archive.IdeaArchive.global.security.GoogleAuthProperties;
+import Idea.Archive.IdeaArchive.global.oauth.GoogleAuthProperties;
 import Idea.Archive.IdeaArchive.global.security.jwt.TokenProvider;
 import Idea.Archive.IdeaArchive.infrastructure.feign.client.GoogleAuth;
 import Idea.Archive.IdeaArchive.infrastructure.feign.client.GoogleInfo;
