+++ src/main/java/Idea/Archive/IdeaArchive/global/oauth/GithubAuthProperties.java
@@ -1,4 +1,4 @@
-package Idea.Archive.IdeaArchive.global.security;
+package Idea.Archive.IdeaArchive.global.oauth;
 
 import lombok.Getter;
 import org.springframework.boot.context.properties.ConfigurationProperties;
@@ -6,15 +6,15 @@ import org.springframework.boot.context.properties.ConstructorBinding;
 
 @Getter
 @ConstructorBinding
-@ConfigurationProperties(prefix = "auth.google")
-public class GoogleAuthProperties {
+@ConfigurationProperties(prefix = "auth.github")
+public class GithubAuthProperties {
 
     private final String baseUrl;
     private final String clientId;
     private final String clientSecret;
     private final String redirectUrl;
 
-    public GoogleAuthProperties(String baseUrl, String clientId, String clientSecret, String redirectUrl) {
+    public GithubAuthProperties(String baseUrl, String clientId, String clientSecret, String redirectUrl) {
         this.baseUrl = baseUrl;
         this.clientId = clientId;
         this.clientSecret = clientSecret;
